backend/requirements.txt
Django==5.1.7
djangorestframework
django-cors-headers
python-dotenv
gunicorn
pillow
djangorestframework-simplejwt
djoser